txpool to mark candidate txs for mining (#772)
as we're now visiting the txpool multiple times during mining we need a
way to ensure we're not trying to include the same transactions twice
and to only get new data each time. This seems a simple way to achieve
this, but if there would be a better way let me know.