Updating sepolia bootnodes (#7314)
Sepolia has had some peering troubles in the past, It seems like the old
EF bootnodes weren't created properly or not added to the IaC stack we
usually have our configs in, so I went ahead and setup new bootnodes on
DigitalOcean. These have more liberal bandwidth limits and cost far
lesser to run, they should be able to help with peering and sync
performance on Sepolia.

The bootnode configs have been commited to the private devops repo along
with a reserved IP and the nodekey has been backed up. The bootnode has
additionally been added to the DevOps monitoring and updating stack. We
should be able to maintain these bootnodes far better in the future.