core: log marshalled receipts upon receipt hash mismatch err in dbg mode (#9340)
Context: we've hit `"mismatched receipt headers for block %d (%s !=
%s)"` a few times for bor consensus, in most cases it's been related to
state sync events. However these sort of errors are difficult to debug
since there's not much visibility to help explain why there is a hash
mismatch with the receipts.

A useful approach is to log the marshalled receipts and compare them to
other functioning nodes (using eth_getBlockReceipts) or blockchain
explorers. This feature is gated behind a dbg env variable