Bor related snapshot changes (#9311)
This PR contains a couple of changes related to bor snapsots:

Its bigger than intended as I used it to produce patch bor snapsots -
and the changes are no difficult to untangle so I want to merge them as
a set.

1. It has some downloader changes which add the following features:

- Added snapshot-lock.json which contains a list of the files/hashes
downloaded - which can be used to manage local state
- Remove version flag and added this to a snapshot type - it has been
used for testing v2 download but is set at v1 for thor PR (see below for
details)
- Manage the state of downloads in the download db - this optimises meta
data look-ups on restart during/after download. For mumbai retrieving
torrent info can take up to 15mins even after download is completed.

2. It has a rationalization of the snapshot processing code to remove
duplicate code between snapshot types and standardize the interfaces to
extract blocks (Dump...) and Index blocks.

- This enables the removal of a separate BorSnapshot and probably
CaplinSnapshot type as the base snapshot code can handle the addition of
new snapshot types.
- Simplifies the addition of new snapshot types (I want to add
borchecploints and bormilestones) as the can be some
- Removes the double iteration from retire blocks
- Aid the insertion of bor validation code on indexing as the common
insertion point is now well defined.

I have tested these changes by syncing mumbai from scratch and by using
it for producing a bor-mainner patch - which starts sync in the middle
of the chain by downloading a previously existing snapshot segment.

I have identified the following issues that I think need to be resolved
before we can use v2 .segs for polygon:

1. For both mumbai and mainnet - downloads are very slow. This looks
like its because lack of peers means that we're hitting the web erver
with many small requests for pieces, which I think the server interpres
as some for of DOS and stops giving us data.
2. Because of the lack of torrents - we can't get metadata - thus don't
start downloading - even if a webpeer is availible.

I'll look to resolve these in the next week or so at which point I can
update the .toml files to include v2 and retest a sync from scratch.