Filter deposit contract logs before pruning (#9123)
This is a hack to not prune logs of certain special contracts,
designated through chain config property, `noPruneContracts`

**Background**
CL needs deposit contract logs to create and maintain its own state
(deposit tree, in particular). It requests the data through
`eth_getLogs`. The naive approach is to not prune any logs of any
contract since the block of the deposit contract. This results in a much
bloated database for Logs. A validating node must be like that.

**Proposed solution**
Add the deposit contract to a special list. On execute stage, skip
pruning Log table. In LogIndex stage, selectively prune logs, i.e. don't
prune those of any of the special contracts' logs.

**Results**
Gnosis pruned node went from about 600GB to 320GB, and can still be a
validating node (not battle tested yet)