Caplin: Implemented Attestation Data production (#9728)
# Caplin: Attestation Production Design

When we are required to produce an attestation data object. we first
asses 2 things:

* What epoch is it from?
* Is it from a future epoch?
* Is it from a past epoch?

This attestationData object changes only within different epochs, it is
extremely cheap to produce `AttestationData` for the current epoch, but
extremely expensive for past and future epochs.

### Current Epoch case

```go
return solid.NewAttestionDataFromParameters(
		slot,                                  // slot requested
		committeeIndex,                        // committee_index request
		headBlockRoot,                         // current head block root
		baseState.CurrentJustifiedCheckpoint(),// current justified checkpoint
		solid.NewCheckpointFromParameters(     // finalization vote
			targetRoot,
			targetEpoch,
		),
	), nil
```

This works in 90 microseconds on an M1 machine so no problem here.

### Past Epoch case

There is no past epoch algorithm for attestation production, what we do
instead is that we keep previously compute `AttestationData` in an
internal cache so that we can just reuse the already computed ones.

### Future Epoch case

Unfortunately, there is no workaround here :(, we just simulate the next
epoch which takes 500ms in case we need to build `AttestationData`s
preemptively.