qa-tests: temporary solution for runner n.1 out of space  (#13245)
tip-tracking and clean-exit (block downloading) require the prebuilt db
to be copied before execution, so as not to corrupt it. But the prebuilt
db has grown in size and there is no room to keep 2 copies of it on the
runner.

So, as a temporary solution, we will run tests on release/2.6x on the
pre-built db itself, and not on its copy.